feat: 게시글 리스트, 상세보기 및 테스트코드

+++ src/domain/board/service/board.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateBoardDTO } from '../dto/create.board.dto';
 import { BoardRepository } from '../repository/board.repository';
 import { UpdateBoardDTO } from '../dto/update.board.dto';
@@ -12,10 +12,40 @@ export class BoardService {
      * 새 글 작성과 기존 글 업데이트를 수행하도록 한다.
      */
     async saveBoard(boardDTO: CreateBoardDTO | UpdateBoardDTO): Promise<BoardEntity> {
-        return await this.boardRepository.save(boardDTO);
+        return this.boardRepository.save(boardDTO);
     }
 
+    /**
+     * 글 번호로 해당 글의 상세정보를 불러오도록 한다.
+     */
+    async readBoard(board_no: number): Promise<BoardEntity> {
+        const boardEntity = await this.boardRepository.read(board_no);
+
+        if (!boardEntity) {
+            throw new NotFoundException('Post not found!');
+        }
+
+        return boardEntity;
+    }
+
+    /**
+     * 글 목록을 불러오도록 한다.
+     * @TODO 페이징, 검색 생각해볼것
+     */
+    async listAllBoard(): Promise<BoardEntity[]> {
+        return this.boardRepository.listAll();
+    }
+
+    /**
+     * 글 번호로 해당 글을 조회해보고 존재하는 글이라면 삭제시키도록 한다.
+     */
     async deleteBoard(board_no: number): Promise<object> {
-        return await this.boardRepository.delete(board_no);
+        const boardEntity = await this.boardRepository.read(board_no);
+
+        if (!boardEntity) {
+            throw new NotFoundException('Post not found!');
+        }
+
+        return this.boardRepository.delete(board_no);
     }
 }
